+++ crates/convergio-db/AGENTS.md
@@ -23,7 +23,7 @@ The block below is rewritten by `cvg docs regenerate` (ADR-0015) —
 do not edit between the markers.
 
 <!-- BEGIN AUTO:crate_stats -->
-**`convergio-db` stats:** 3 `*.rs` files / 8 public items / 240 lines (under `src/`).
+**`convergio-db` stats:** 3 `*.rs` files / 8 public items / 242 lines (under `src/`).
 
 No files within 50 lines of the 300-line cap.
 <!-- END AUTO -->
